This commit is a candidate for alpha testing as a downloadable app at GitHub.  Hopefully, it will be available under the Releases tab soon.
Code-signing with a registered Apple Developer ID and sandboxing are intended to be working now, but those are not independently verified yet.

An issue exists where the LocalRadio app cannot launch the AudioMonitor tool if the app is located within a folder containing spaces in the folder name.  The user will be alerted at launch time if the problem is detected.

The build phases were modified in the main project and subproject.