For next LocalRadio v1.2-alpha release -
Some bugs were fixed in the menus for "Audio Output" and "Stream Source", where Core Audio devices were not configured correctly for Sox.

Typically, both of those items can be set to "Built-in Icecast Server" in order to support use of the web interface on iPhone, etc.  If needed, the Audio Output can also be set to one of the Mac's audio devices.

Some virtual audio devices can also be specified for favorite frequencies and categories, for example, the Soundflower system extension (https://github.com/mattingalls/Soundflower/releases/tag/2.0b2).  This can be useful with some digital audio decoder apps that use Core Audio for input and output.  LocalRadio can send the demodulated audio from the radio to a Soundflower virtual audio device for decoding, and then receive the decoded audio from a different Soundflower device and send it to the Icecast output tasks for streaming.

Apple's "Audio MIDI Setup" app is also useful for managing audio devices.  It is located in the Mac's /Applications/Utilities/ folder.