The version is set to LocalRadio v1.4a.
A new option was added to the "Change Configuration…" window, a checkbox button named "Use Auto-Play in this app".  It only affects the web interface in the LocalRadio application, not remote clients or the direct audio from AudioMonitor if that mode is used.  Currently, the default state for the checkbox when the app launches is NO.

New web interface application icons were added for iPads.

The SVG icons are now inserted dynamically into the html document instead of using <object> tags, and CSS rules added for scaling the icons for mobile devices.  The SVG files are still separate documents that can be edited as needed.

On the web interface tuner pages where the user enters the frequency by clicking on a digit, then clicking on up arrow and down arrow buttons, the web browser will use the speech synthesizer API to speak the frequency.

In the web interface, a "title" attribute was added to several visual elements.  These will be displayed as tooltips when the mouse hovers over the titled item.

A possible bug fix was flagged in some original rtl_fm code where ss_busy was referenced in a line of code - unassigned - right below "worst case, nuke a buffer".  I assigned ss_busy to 1.