In WebServerConnection generateAudioPlayerString, the code to append a random query to the URL in order to acquire a fresh streaming connection from Icecast is disabled for now.  It was causing problems with unnecessary streaming stalls in some remote web browsers, and the main app itself.
Further testing is needed to determine if this change causes a situation where the HTML audio object is paused by the user, but continues to download and buffer the audio stream.  The goal is for the remote players to get the closest thing to live radio as possible, without unnecessary buffering.

Also in WebServerConnection, some NSLog() instrumentation is temporarily added to log HTTP requests for HTML data, and can be viewed with Console.app.  This will be helpful in finding out why some requests are slow to respond.