In various places where the MP3 encoder bitrate setting is managed, the setting is restricted to valid bitrates defined for the LAME library, which is used by Sox as the current MP3 encoder.  Some of the previous settings were invalid (like 3 Kbps), and due to several bugs, in all cases LocalRadio was producing a 320 Kbps stream.  The new settings should be much more accurate, and produce streams ranging from 32 Kbps to 320 Kbps.  Some cosmetic work remains to be done, like setting the current bitrate in a couple of dropdown menus.
A new button labeled "Open Audio Player Page" is now available in the "Now Playing" page in the web interface.  Click or tap that button to load a new standalone audio player page.  (If using the LocalRadio web view, right-click in the page and select the "Back" command to exit the standalone player.  This will be improved in a future commit.)

For radio signals with a low sampling rate, despite some resampling stages in the processing pipeline, the HTML audio element in Safari and the Cocoa WebView have a problem with the stream stalling periodically during playback, rebuffering, and resuming after a few seconds of silence.  Other streaming audio clients like Chrome for Android and Mac, QuickTime Player and VNC play the same audio streams smoothly without stalling.  Some future work is planned for the audio processing pipeline that might fix the problem in the future, but a workaround is available now: use the "Open Audio Player Page" button described above.