The processing pipeline is slightly revised in this commit.  The AudioMonitor object uses the same stdin data from rtl_fm_localradio, but the pass-through data is now always resampled to 48000 Hz before it is written to stdout.  Before this change, the next stage in the pipeline, a Sox process, was used to resample the data.  Sox no longer does the resampling since AudioMonitor has already performed that task.  However, the Sox process remains in place to apply other filters, like deemphasis.
A few audio packets are still getting dropped periodically, like in a batch of several dozen SInt16 samples every second, but the audio seems to sound okay anyway.

More good news: moving the 48000 Hz resampling process to an earlier stage seems to fix the MP3 player stream stall problem with the <audio> element with Safari and Cocoa WebView on Mac and iOS.  Most features seem to be running smoother now.

The version was incremented to 1.12-alpha.