In TaskPipelineManager, a new kTaskPipelineStatusTerminating status code was added between the kTaskPipelineStatusRunning and kTaskPipelineStatusTerminated phases.
In TaskPipelineManager terminateTasks, the @synchronized directive was used in a critical section of code, and the result of NSTask isRunning is checked before sending the terminate message.  This should avoid some crashes that occurred when terminate was sent to an NSTask that had already terminated.

Also, the NSTask terminate section of code is now enclosed in a @try/@catch block to avoid app crashes.