The new feature for using a Mac Core Audio input device is implemented in this commit.  It is selected by clicking the Devices item in the top-level menu.  A list of available audio input devices can be selected from a pop-up menu, then click the Listen button to start streaming.
A change was made in the audio processing pipeline.  The task for user-specified SoX processing is created only if the user actually specified some sox processing.  Otherwise, that sox task is omitted from the pipeline; in that scenario, the AudioMonitor task will send piped output directly to the UDPSender task.