In AppDelegate, the updateCurrentTasksText method will include the icecastStatusDictionary obtained from the Icecast server at /admin/stats .
In EZStreamController and IcecastController, the server config files for Icecast and EZStream are changed to use the Icecast password generated by the app.

In IcecastController, the icecastStatusDictionary method sets a flag when an update query is in progress.  If an update is in progress when icecastStatusDictionary is called, a copy of the most recent dictionary is returned.

In WebServerConnection, some instrumentation was added to log requests and file responses.