In AppDelegate updateCurrentTasksText:, exit the method immediately if the application is terminating.  This will avoid a potential crash when LocalRadio.app is quitting.
In Info.plist, some changes were made in the record for NSAppTransportSecurity to acommodate the current Icecast non-TLS transport configuration (http, not https).

In localradio.js, the postMessage("startaudio", "*") calls in four methods were relocated to get called immediately after the "listen" request is sent, rather than getting called after the response for the "listen" request is received.  It has to be this way due to Apple's policy on starting audio.

In WebViewDelegate, a webView:decidePolicyForNavigationAction:request:frame:decisionListener: method was added, which simply calls [listener use].