This commit features a new option for FM stereo radio listening, with the addition of two open source tools -
wmf-tools by Oona Räisänen - https://github.com/windytan/wfm-tools

liquid-dsp by Joseph D. Gaeddert - https://github.com/jgaeddert/liquid-dsp

A sample rate of at least 106000 is required for stereo usage.  The stereo option usually works best with a station with a strong signal, with the gain reduced to a value like 25.4 or 12.5.

The addition of stereo required several changes, especially in the AudioMonitor unit.  AudioMonitor can now accept input with either one or two channels, and it always outputs two channels at 48000 Hz.  All subsequent stages in the processing pipeline then use two channels for input and output.

A new "stereo_flag" column was added to the "frequency" table in the LocalRadio database.  The web user interfaces have been changed to allow the user to enable or disable the stereo option.

A couple of minor memory leaks were fixed.

The web interface button in the Settings page to restart the server was disabled in the previous release.  The button works now.

The LocalRadio SVG icon was added to the web documents, and is shown in the Info page.