Added a Commands menu with command to Show FCC Search Window and Reload Web View.
This commit preliminary support for custom tasks that can be used in place of LocalRadio's version of rtl_fm.  This allows external command-line tools to be used as an audio source, if they output LPCM data to stdout.  If LocalRadio is built with Apple's sandbox option, the external tool must be code-signed with the same certificate as LocalRadio.  Prior to code-signing, the external tool should be inspected with "otool -L myToolName" to see the load paths for any library dependencies, and modified with "install_name_tool" to fix the locations of non-system libraries.  The package of code-signed tools in the LocalRadio Frameworks bundle can be used by the external tool.  Otherwise, any library dependencies may also need to be copied, modified with install_name_tool and code-signed.   Custom tasks are stored in a new table named "custom_task".  A user interface design is partially implemented, but not ready for use.