The LocalRadio database in the Applications Support folder now has a new file name: "LocalRadio-V2.sqlite3".  When the app launches, if a database file with the new name is not found, the app will create a new V2 database, and attempt to import records from a database with the old file name "LocalRadio.sqlite3".  If an old database file is not found, the database will be initialized with a few sample records from "data_skeleton.sqlite3".  A couple of columns were removed in the new database's frequency table, and a stereo_flag column was added.  (The "local_radio_config" table is currently not imported, so the default settings from "data_skeleton.sqlite3" are used).
Implementation of the new Custom Tasks feature is mostly complete now.  An example of a custom task using the bundled rtl_fm_localradio tool is included when at new database is created.  External executables can also used instead of rtl_fm_localradio in the sandboxed build of LocalRadio if the external executable is processed with the codesign tool, using the same certificate used to sign LocalRadio.app (and bundled tools like rtl_fm_localradio).  To use a defined Custom Task, navigate to the "Devices" page and use the controls there.

An alternative to code-signing an external executable is to build the main app and all of the libraries and bundled tools with sandboxing disabled.  Disabling the app sandbox will result in more manual interventions to run the system, to allow icecast to accept incoming connections while the macOS firewall is active, for example.