This commit has a temporary fix for custom tasks, by always using SoX in the processing chain.  This is likely due to an output buffering issue in AudioConverter2.  Ideally, SoX should not be required unless the user specified a SoX filter.
The AudioConverter for HE-AAC at 32000 bitrate is still not working, but AAC-LC seems to be working okay at 64000 and 128000.  This commit contains some changes for audio MIME type for the web page and the Icecast server.

A bug was fixed for icecastMountName.

In TaskPipelineManager, the first task now has stdin set to a null device rather than an empty pipe.