This commit adds support for Icecast streaming with SSL/TLS encryption.  Icecast is configured with two ports to stream LocalRadio audio, via http and https schemes.
The source audio connection uses http to stream to the Icecast server.  Since the Icecast server and the source are both running on the same hosts, there was little benefit to encrypting that connection.

Although the Icecast server now supports https, the user interface web server currently still supports only http.  The plan is to add https support there in a future commit.

The Bonjour hostname (e.g., joes-macintosh.local) is now used to identify the host in most cases.

An experimental new tool was created - IcecastSourceClient - and a customized version libshout library to support AAC was also added.  It was originally hoped that libshout would provide an easier way to encrypt the source connection, but I never got it to work due to various issues with the app sandbox and BoringSSL.  The sandbox problems were mostly resolved by removing the MacPorts versions of OpenSSL's libssl and libcrypto which needed file access (for root certificates, etc.), and libcurl, libxml and libxslt, and instead using the standard Mac versions of those libraries in /usr/lib/, etc.  Some code units were inspected and modified with otool and install_name_tool to use the Mac libraries.  Ultimately, the previous "IcecastSource" tool turned out to work better than "IcecastSourceClient" with libshout, so the app is sticking with IcecastSource.  The unused libshout source will remain in the project for reference purposes for now.

A bug was fixed where the main thread was using a sleep command.  A bug was fixed to avoid a crash during status updates.

Some fflush() commands were added to stdout functions in AACEncoder, AudioMonitor2.