In some Xcode schemes, the CFNETWORK_DIAGNOSTICS environment variable was set.
In AudioMonitor2, some adjustments were made in the main input loop to produce larger packets for AAC encoding in the subsequent process.

In TPCircularBuffer.h, some diagnostic information is output to stderr if a fatal error condition is found.

In IcecastSource main.m, some minor reformatting was done.

In AppDelegate, the Bonjour service names for the http and https servers were made distinct - "LocalRadio" and "LocalRadioSecure".  When the application terminates, the Bonjour service names will be removed from the list of current services.

In HTTPWebServerConnection, the "Now Playing" menu will display the current frequency for stations that are selected with the manual tuners.  For stations selected in Favorites or Categories, the user's specified title will continue to be displayed.  Also, a typo was fixed in the general tuner display for Now Playing, instead of "Untitled FM Frequency".  In generateAudioPlayerString(), for the HTTP server an IP address will be returned, and for HTTPS server the Bonjour host name will be returned.  A similar change was applied to generateOpenAudioPlayerPageButtonString();

In HTTPWebServerController, the serviceName method was corrected to "LocalRadio" instead of "LocalRadioSecure".

In icecast.xml, the loglevel value was reduced from 4 (debug) to 3 (info).

In IcecastController's icecastStatusDictionary method, the URL request timeout interval value was increased to 0.1 to 0.5.

In Info.plist, the ITSAppUsesNonExemptEncryption key was added.

In LocalRadioAPI, requests were changed to use HTTPS instead of HTTP.

In index.html, custom.css and localradio.js, some changes were made to try to prevent page zooming on iOS devices when the +/- frequency setting buttons are tapped - to no avail!

In index.html's startAudioPlayerAfterDelay() function, the call to seekToEndForAudioElement() is made only if the audio buffer contains some future data to play.  Also, a workable delay function was found, since setTimeout() is not allowed to call play(), etc. due to Apple's rules.  In updateStatusDisplay(), the JSON.parse() function is now called from within a try/catch structure.  That function is prone to failure if some invalid JSON data was received.

In UDPListener, a new property named accumulatedAudioData was added.  It is used to increase the packet size for the subsequent AACEncoder process.