In TPCircularBuffer.h, some error message were added for stderr output to gather information about why it fails sometimes.
In the main window under the LocalRadio tab, the URLs for both the secure HTTPS and nonsecure HTTP web servers are displayed, and can be opened is an external web browser by licking on the URL text.  AppDelegate and URLTextField were modified accordingly.

In TLSManager, animation is started and stopped for the progress indicator when keys and certificates are being generated.