In AudioMonitor2, some parameters for usleep() were changed, and some buffer size calculations were changed.  Some dead code was removed.
In SDRController, the SoX task item will be created only if the audioOutputFilterString is not empty.

In UDPListener, a usleep() time value was changed.

In UDPSender, the maximum output data size to stdout was changed from 4096 to 2048.  That value seems to work more smoothly with the recently added fflush() call.