This commit discontinues the use of pre-built dylib files that were copied from MacPorts and statically stored in the project.  Those packages are now deleted from the project.  It is now assumed that MacPorts is installed on the build machine, and all of the packages required to build LocalRadio are installed in MacPorts.
A new build phase is added to run the import_dylibs2.sh script.  In that bash script, the "TARGETS" variable lists the dylibs that are required.  The script will not run MacPorts to install the packages, it assumes the user has done so already in a command-line session, using "sudo port install <package-name>" for each required package.

In the Xcode project navigator, the contents of the "Libraries_Modified" folder should match the TARGETS list in the import_dylibs2.sh script.

Additionally, the "dylibbundler" MacPorts package should be called for use in import_dylibs2.sh.

The script will copy the specified dylibs into a folder titled  "Libraries_Modified" within the build products folder.  If that folder exists, the script will assume that build phase has successfully executed in a previous build and exits immediately.  If a problem occurs when the script runs, the Libraries_Modified folder should be deleted before the next build in order to try again.

There are some advantages to this approach, including improved confidence in code security by using fresh libraries from MacPorts instead of relying on prebuilt-libraries that were used previously.  And it should make it easier to build the application for older versions of Mac OS X.