This commit adds Icecast-Server as a git submodule.  The goal is to build the icecast server with custom options to remove dependencies on some unnecessary libraries, specifically: --without-speex and --without-curl
Some of the remaining library dependencies will be able to use the standard macOS libraries instead of MacPorts.  It is especially helpful for OpenSSL to use the macOS library in order to function correctly in the sandbox environment.

Instructions for building the submodule will be contained in a new file - "Icecast-Server git and build notes.txt"