More changes for Catalina -
The HTTPS streaming audio with a Bonjour host name does not work in Catalina or iOS 13, so that feature has been removed as a user interface option.

For the remaining HTTP connection, a Sharing button has been added to provide a convenient way to open LocalRadio on an iPhone via AirDrop.

In a few places, the Xcode static analyzer found some secondary processes trying to call methods that are allowed on the main thread only.  Those lines of code were wrapped in GCD calls to switch to the main thread when needed.