In HTTPWebServerConnection, for using Audio Devices input, a text field was added to the HTML form where the user can specify a sox filter.  The default value is "vol 1".  Typically, the user might want to adjust the volume to 2 or more.
In SDRController, use the new sox filter string in the first stage chain, in the final sox task.  Also, obtain the current sample rate for the audio input device and use it in the sox task that captures the audio input, typically 48000 or 44100.  Previously, the standard macOS "Audio MIDI Setup.app" application could be used to adjust the device sample rate to 48000 before using LocalRadio to generate the audio stream, but now LocalRadio will obtain the correct value and use it automatically.